requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
pandas==2.1.4
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.0.2